requirements.txt
# numpy==1.19.5
opencv-python
# pkg_resources==0.0.0
aioredis
asgiref
async-timeout
asyncpg
attrs
bcrypt
certifi
cffi
charset-normalizer
click
databases
fastapi
h11
hiredis
idna
iniconfig
numpy==1.21.1
packaging
passlib
pluggy
psycopg2-binary
py
pycparser
pydantic
PyJWT
pyparsing
pytest
python-multipart
requests
six
SQLAlchemy
starlette
toml
typing-extensions
urllib3
uvicorn
pillow